CI - refactor to use podman and single job for multi platform build
The GitHub Workflow to do do a matrix build of multiple platforms and then
bring them togeather into a single manifest is annoyingly complicated and
hard to follow.

From testing the matrix jobs and the do it all at once, at least for
two platforms, is the same time (~3 minutes).

This commit simplies for the worklow code at expense of not seeing
the two platform images being built in GitHub actions.